+++ src/components/ProductTile/ProductTilePrice.tsx
@@ -5,6 +5,8 @@ import './style/ProductTilePrice.less';
 import AnimationValue from './ProductTileValue';
 
 interface IProductTilePriceProps {
+    /** title */
+    title?: string;
     /** value */
     value: string;
     /** unitExtra */
@@ -18,6 +20,7 @@ interface IProductTilePriceProps {
 const cn = cnCreate('mfui-product-tile-price');
 class ProductTilePrice extends React.Component<IProductTilePriceProps> {
     static propTypes = {
+        title: PropTypes.string,
         value: PropTypes.string.isRequired,
         unitExtra: PropTypes.string,
         unitValue: PropTypes.string,
@@ -25,13 +28,13 @@ class ProductTilePrice extends React.Component<IProductTilePriceProps> {
     };
 
     render() {
-        const { value, unitExtra, unitValue, discount } = this.props;
+        const { title, value, unitExtra, unitValue, discount } = this.props;
         const valueElem = <AnimationValue value={(discount || value) + ` ${unitValue}`} />;
         const unitElem = <span className={cn('payment-period')}>{unitExtra}</span>;
 
         return (
             <div className={cn('', { discount: !!discount })}>
-                <div className={cn('discount-condition')}>При покупке новой SIM–карты</div>
+                <div className={cn('discount-condition')}>{title}</div>
                 <span className={cn('old-price')}>
                     <span className={cn('old-price-value')}>
                         {discount ? `${value} ${unitValue}` : ' '}